+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## unreleased
+* Add `getBillingAddress` method to `PaymentMethodNonceDetailsPayerInfo`
+* Add `getShippingAddress` method to `PaymentMethodNonceDetailsPayerInfo`
+
 ## 2.108.0
 * Fix null pointer exception when 3DS lookup returns a `422`
 * Add ThreeDSecure test payment method nonces